+++ src/extra/gfx/feature.cpp
@@ -50,7 +50,6 @@ ENUM_HELP(gfx::BuiltinFeatureId, gfx::BuiltinFeatureId::Velocity,
           SHCCSTR("Outputs object velocity into the velocity global & output"));
 
 namespace gfx {
-using shards::Mat4;
 
 struct BuiltinFeatureShard {
   DECL_ENUM_INFO(BuiltinFeatureId, BuiltinFeatureId, 'feid');
@@ -240,8 +239,9 @@ struct FeatureShard {
   }
 
   void cleanup() {
-    for (auto &pair : _generatorCapturedVariables)
+    for (auto &pair : _generatorCapturedVariables) {
       releaseVariable(pair.second);
+    }
     _generatorCapturedVariables.clear();
 
     _viewGeneratorsMesh.reset();
@@ -546,8 +546,7 @@ struct FeatureShard {
       v = *ref;
     }
 
-    ParamVariant variant;
-    varToParam(v, variant);
+    ParamVariant variant = varToParam(v);
 
     if (ref)
       releaseVariable(ref);
@@ -660,12 +659,15 @@ struct FeatureShard {
   }
 
   void addGeneratorCapturedVariablesToMeshes() {
-    for (auto &pair : _generatorCapturedVariables) {
-      if (_viewGeneratorsMesh)
-        _viewGeneratorsMesh->variables.emplace(pair.first, *pair.second);
-      if (_drawableGeneratorsMesh)
-        _drawableGeneratorsMesh->variables.emplace(pair.first, *pair.second);
-    }
+    auto forEachMesh = [&](auto mesh) {
+      for (auto &pair : _generatorCapturedVariables) {
+        if (mesh) {
+          mesh->refs.emplace(pair.first, pair.second);
+        }
+      }
+    };
+    forEachMesh(_viewGeneratorsMesh);
+    forEachMesh(_drawableGeneratorsMesh);
   }
 
   SHVar activate(SHContext *context, const SHVar &input) {
@@ -777,6 +779,7 @@ struct FeatureShard {
         mesh->schedule(wire, input, false);
       } else {
         // Update inputs
+        // NOTE: this is owned by the wire
         (TableVar &)wire->currentInput = std::move(input);
       }
     }
